cleaned up code; added make hifidata and instructions
