+++ pytest.ini
@@ -2,4 +2,3 @@
 markers =
     bandage: bandage required tests
     long: long running tests
-    extralong: extra long running tests
