+++ yeat/tests/data/pacbio.cfg
@@ -1,0 +1,10 @@
+[
+    {
+        "algorithm": "canu",
+        "extra_args": "genomeSize=4.8m"
+    },
+    {
+        "algorithm": "flye",
+        "extra_args": ""
+    }
+]
